+++ src/node-lib/curriculum-api-2023/queries/specialistLessonDownload/specialistLessonDownload.query.ts
@@ -2,6 +2,7 @@ import { z } from "zod";
 
 import { Sdk } from "../../sdk";
 import { lessonDownloadsListSchema } from "../../shared.schema";
+import { SpecialistLessonDataRaw } from "../specialistLessonOverview/specialistLessonOverview.schema";
 
 import {
   SpecialistLessonDownloadRaw,
@@ -80,14 +81,16 @@ export const constructDownloadsArray = (
 };
 
 export const constructHasDownloadableResources = (
-  lesson: SpecialistLessonDownloadRaw,
+  lesson: SpecialistLessonDownloadRaw | SpecialistLessonDataRaw[number],
 ) => {
   return (
     (!!lesson.presentation_url &&
       lesson.contains_copyright_content === false) ||
     (!!lesson.starter_quiz && !!lesson.starter_quiz_asset_object) ||
     (!!lesson.exit_quiz && !!lesson.exit_quiz_asset_object) ||
-    !!lesson.worksheet_url
+    (!!lesson.worksheet_url &&
+      typeof lesson.worksheet_asset_object
+        ?.google_drive_downloadable_version === "string")
   );
 };